requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
django-seed==0.1.9
djangorestframework==3.8.2
Faker==0.8.15
Markdown==2.6.11
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
text-unidecode==1.2
